+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+- More fixes for disconnected clients appearing in Trace dropdown
+
+## [1.2.1] - 2025-10-13
+
+### Added
+
 - Support for `@disconnect` / `@dc` in `Send` packet syntax
 - Client `DismissParcel` command (only found on older clients)